Code refactor
1. Refactor message struct and payload models
2. Remove unused module
3. Added input_topic to workflow
