+++ core-data/src/models/message/errors.rs
