+++ core-data/src/models/message/mod.rs
@@ -1,0 +1,15 @@
+mod core;
+mod payload;
+mod auditlog;
+mod progress;
+mod parse;
+mod enrich;
+
+mod errors;
+mod iso20022;
+
+pub use self::core::Message;
+pub use self::payload::{Payload, PayloadFormat, PayloadSchema, Encoding, StorageType};
+pub use self::auditlog::{AuditLog, ChangeLog};
+pub use self::progress::{Progress, MessageStatus, StatusCode};
+pub use self::enrich::EnrichmentConfig;
