+++ core-data/src/models/message/progress.rs
@@ -1,0 +1,31 @@
+use serde::{Deserialize, Serialize};
+use time::OffsetDateTime;
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+pub struct Progress {
+    pub status: MessageStatus,
+
+    pub workflow_id: String,
+
+    pub prev_task: String,
+    
+    pub prev_status_code: Option<StatusCode>,
+    
+    #[serde(with = "time::serde::iso8601")]
+    pub timestamp: OffsetDateTime,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+pub enum MessageStatus {
+    Recieved,
+    Processing,
+    Completed,
+    Failed,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+pub enum StatusCode {
+    Success,
+    Failure,
+}
+
