+++ core-data/src/models/mod.rs
@@ -1,8 +1,4 @@
-pub mod payload;
-pub mod auditlog;
 pub mod message;
 
 pub mod task;
 pub mod workflow;
-pub mod errors;
-pub mod iso20022;
